requirements.txt
# FastAPI and ASGI server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.7

# Database (MODERN)
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Authentication and Security (MODERN)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.4
pydantic-settings==2.3.0  # Replaces old way of handling settings

# HTTP Client
aiohttp==3.9.5
httpx==0.27.0

# WebSocket support
websockets==12.0

# Background tasks
celery==5.4.0
redis==5.0.4

# File handling
openpyxl==3.1.2
pandas==2.2.2
aiofiles==23.2.1

# Email
aiosmtplib==3.0.1
jinja2==3.1.4

# Logging
structlog==24.1.0

# Monitoring and metrics
prometheus-client==0.20.0

# Development and testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# Code quality
black==24.4.2
isort==5.13.2
flake8==7.1.0
mypy==1.10.0

# Documentation
mkdocs==1.6.0
mkdocs-material==9.5.26

# Utilities
python-dateutil==2.9.0
pytz==2024.1
click==8.1.7
rich==13.7.1

# Telnet client (for Jasmin integration)
telnetlib3==2.0.4

# Image processing
Pillow==10.3.0

# Validation
email-validator==2.1.1
phonenumbers==8.13.37

# Caching
aiocache==0.12.2

# Rate limiting
slowapi==0.1.9

# Environment variables
python-dotenv==1.0.1

# JSON handling
orjson==3.10.3

# Async utilities
asyncio-mqtt==0.16.1

# Health checks (Use FastAPI's own middleware or a modern library if needed)
# fastapi-health REMOVED - often not needed with modern FastAPI

# CORS handling
# fastapi-cors REMOVED - integrated into FastAPI, use CORSMiddleware directly

# Pagination
fastapi-pagination==0.12.23

# File storage
boto3==1.34.113

# System monitoring
psutil==5.9.8

# Network utilities
netifaces==0.11.0

# JWT handling
pyjwt==2.8.0

# ---- PACKAGES REMOVED DUE TO INCOMPATIBILITY ----
# databases - Not compatible with SQLAlchemy 2.0
# fastapi-websocket-pubsub - Requires Pydantic v1 and old websockets
# fastapi-users - The old version has complex dependencies. Use a modern version if needed, or implement auth directly.
# http-status - Not a standard or necessary package.
